requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]==2.5.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
groq==0.4.1
pandas==2.1.3
openpyxl==3.1.2
websockets==12.0
slowapi==0.1.9
argon2-cffi==23.1.0